updater/requirements.txt
pandas
psycopg[binary]
PyYAML
yfinance
